+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/festival/FestivalDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package com.daedan.festabook.data.datasource.remote.organization
+package com.daedan.festabook.data.datasource.remote.festival
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
 import com.daedan.festabook.data.model.response.festival.FestivalResponse
