+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/festival/FestivalNotificationDataSource.kt
@@ -1,9 +1,9 @@
-package com.daedan.festabook.data.datasource.remote.organization
+package com.daedan.festabook.data.datasource.remote.festival
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
 import com.daedan.festabook.data.model.response.festival.FestivalNotificationResponse
 
-interface OrganizationBookmarkDataSource {
+interface FestivalNotificationDataSource {
     suspend fun saveOrganizationBookmark(
         organizationId: Long,
         deviceId: Long,
