+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/festival/FestivalNotificationDataSourceImpl.kt
@@ -1,13 +1,13 @@
-package com.daedan.festabook.data.datasource.remote.organization
+package com.daedan.festabook.data.datasource.remote.festival
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
 import com.daedan.festabook.data.model.request.FestivalNotificationRequest
 import com.daedan.festabook.data.model.response.festival.FestivalNotificationResponse
 import com.daedan.festabook.data.service.OrganizationBookmarkService
 
-class OrganizationBookmarkDataSourceImpl(
+class FestivalNotificationDataSourceImpl(
     private val organizationBookmarkService: OrganizationBookmarkService,
-) : OrganizationBookmarkDataSource {
+) : FestivalNotificationDataSource {
     override suspend fun saveOrganizationBookmark(
         organizationId: Long,
         deviceId: Long,
